# api/openapi.yaml
openapi: 3.0.4
info:
  title: Shoplist API
  description: |-
    Official OpenAPI specification for the Shoplist API

    Some useful links:
    - [Swagger Editor](https://editor.swagger.io/)
  contact:
    url: https://github.com/jkutkut/shoplist
  license:
    name: GPL 3.0
    url: https://github.com/jkutkut/shoplist/blob/main/LICENSE
  version: 1.7.0
servers:
  - url: http://localhost/api
paths:
  /v1/user/{user_id}:
    get:
      summary: Get user info
      description: Get user info by user ID
      tags:
        - user
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/uuid'
      responses:
        '200':
          description: User info returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
    delete:
      summary: Delete user
      description: Delete user
      tags:
        - user
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/uuid'
      responses:
        '200':
          description: User deleted
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
...
  /v1/user/me:
    get:
      summary: Get current user info
      description: Get current user info based on JWT token
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
  /v1/user/me/token:
    post:
      summary: Refresh JWT token
      description: Refresh JWT token for current user by invalidating the current token and returning a new one
      tags:
        - user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token refreshed, new JWT token returned
          headers:
            set-cookie:
              schema:
                $ref: '#/components/schemas/bearer-token'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    $ref: '#/components/schemas/jwt'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
...
  /v1/user/superuser/{user_id}:
    post:
      summary: Set user as superuser
      description: Set user as superuser by user ID. Requires superuser permissions.
      tags:
        - user
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/uuid'
      responses:
        '200':
          description: User set as superuser
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
    delete:
      summary: Remove superuser from user
      description: Remove superuser from user by user ID. Requires superuser permissions.
      tags:
        - user
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/uuid'
      responses:
        '200':
          description: Superuser removed from user
        '401':
          description: Invalid credentials, user not found, not enough permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/result'
...
components:
...
  schemas:
...
    jwt:
      type: string
      description: JWT token
    bearer-token:
      type: string
      description: Bearer token
      example: Bearer <jwt token>
    # Common
    uuid:
      type: string
      format: uuid
      description: A 128-bit number encoded as a string.
      minLength: 36
      maxLength: 36
    name:
      type: string
      example: Marvin
...
    image:
      type: string
      nullable: true
      description: A URL to an image representing the entity.
    created_at:
      type: string
      format: date-time
      description: The date and time when the entity was created.
      example: 1970-01-01T00:00:00Z
    updated_at:
      type: string
      format: date-time
      description: The date and time when the entity was last updated.
      example: 1970-01-01T00:00:00Z
...
    result:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: The HTTP status code.
          example: 401
        message:
          type: string
          description: A message describing the error.
          example: Invalid credentials
...
    user:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/uuid'
        name:
          $ref: '#/components/schemas/name'
        image:
          $ref: '#/components/schemas/image'
        created_at:
          $ref: '#/components/schemas/created_at'
        updated_at:
          $ref: '#/components/schemas/updated_at'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT